Macro claim::assert_ready

macro_rules! assert_ready {
    ($cond : expr,) => { ... };
    ($cond : expr) => { ... };
    ($cond : expr, $($arg : tt) +) => { ... };
}

Asserts that expression returns Poll::Ready(T) variant.

This macro is available for Rust 1.36+.

Uses

Assertions are always checked in both debug and release builds, and cannot be disabled. See debug_assert_ready! for assertions that are not enabled in release builds by default.

Custom messages

This macro has a second form, where a custom panic message can be provided with or without arguments for formatting. See std::fmt for syntax for this form.

Examples

let res = Poll::Ready(42);

assert_ready!(res);

Value of T type from the Poll::Ready(T) will also be returned from this macro call:

let res = Poll::Ready(42);

let value = assert_ready!(res);
assert_eq!(value, 42);

Poll::Pending variant will cause panic:

let res = Poll::Pending;

assert_ready!(res);  // Will panic